Refactor components files
