+++ client/src/components/coin/TokenPageDivider.js
@@ -1,0 +1,9 @@
+import { Center, Divider } from "@chakra-ui/react";
+
+export default function TokenPageDivider() {
+  return (
+    <Center height="50px">
+      <Divider m={"10px"} orientation="vertical" borderColor="gray.500" />
+    </Center>
+  );
+}
